Move types to a new file
