+++ src/main/kotlin/org/shardscript/semantics/core/Symbol.kt
@@ -7,8 +7,6 @@ import org.shardscript.semantics.infer.SubstitutionChain
  */
 sealed class Symbol
 
-sealed interface Type
-
 sealed class SymbolTableElement: Symbol() {
     abstract val parent: Scope<Symbol>
 }
@@ -22,8 +20,6 @@ sealed class NamedSymbolWithMembers(
     private val symbolTable: SymbolTable = SymbolTable(parent)
 ): NamedSymbolTableElement(), Scope<Symbol> by symbolTable
 
-data object ErrorSymbol : Symbol(), Type
-
 data class Block(
     override val parent: Scope<Symbol>,
     private val symbolTable: SymbolTable = SymbolTable(parent)
@@ -36,14 +32,6 @@ data class LocalVariableSymbol(
     val mutable: Boolean
 ) : NamedSymbolTableElement()
 
-/**
- * Function Primitives
- */
-data class FunctionTypeSymbol(
-    val formalParamTypes: List<Type>,
-    val returnType: Type
-) : Symbol(), Type
-
 data class FunctionFormalParameterSymbol(
     override val parent: Scope<Symbol>,
     override val identifier: Identifier,
@@ -52,74 +40,6 @@ data class FunctionFormalParameterSymbol(
     var costMultiplier: CostExpression = CommonCostExpressions.defaultMultiplier
 }
 
-/**
- * Type/Fin Primitives
- */
-sealed class TypeParameter : Symbol(), Type
-
-data class StandardTypeParameter(
-    val qualifiedName: String,
-    val identifier: Identifier
-) : TypeParameter(), Type
-
-data class ImmutableFinTypeParameter(
-    val qualifiedName: String,
-    val identifier: Identifier
-) : TypeParameter(), CostExpression, Type {
-    override fun <R> accept(visitor: CostExpressionVisitor<R>): R {
-        return visitor.visit(this)
-    }
-}
-
-data class MutableFinTypeParameter(
-    val qualifiedName: String,
-    val identifier: Identifier
-) : TypeParameter(), CostExpression, Type {
-    override fun <R> accept(visitor: CostExpressionVisitor<R>): R {
-        return visitor.visit(this)
-    }
-}
-
-data class FinTypeSymbol(val magnitude: Long) : Symbol(), CostExpression, Type {
-    override fun <R> accept(visitor: CostExpressionVisitor<R>): R {
-        return visitor.visit(this)
-    }
-}
-
-data object ConstantFinTypeSymbol : Symbol(), CostExpression, Type {
-    override fun <R> accept(visitor: CostExpressionVisitor<R>): R {
-        return visitor.visit(this)
-    }
-}
-
-data class SumCostExpression(val children: List<CostExpression>) : Symbol(), CostExpression {
-    override fun <R> accept(visitor: CostExpressionVisitor<R>): R {
-        return visitor.visit(this)
-    }
-}
-
-data class ProductCostExpression(val children: List<CostExpression>) : Symbol(), CostExpression {
-    override fun <R> accept(visitor: CostExpressionVisitor<R>): R {
-        return visitor.visit(this)
-    }
-}
-
-data class MaxCostExpression(val children: List<CostExpression>) : Symbol(), CostExpression {
-    override fun <R> accept(visitor: CostExpressionVisitor<R>): R {
-        return visitor.visit(this)
-    }
-}
-
-/**
- * Generic Primitives
- */
-data class SymbolInstantiation(
-    val substitutionChain: SubstitutionChain
-) : Symbol(), Type
-
-/**
- * Function Types
- */
 data class GroundFunctionSymbol(
     override val parent: Scope<Symbol>,
     override val identifier: Identifier,
@@ -159,10 +79,6 @@ data class ParameterizedFunctionSymbol(
     fun type() = FunctionTypeSymbol(formalParams.map { it.ofTypeSymbol }, returnType)
 }
 
-/**
- * Data Type Primitives
- */
-
 data class FieldSymbol(
     override val parent: Scope<Symbol>,
     override val identifier: Identifier,
@@ -175,59 +91,6 @@ data class PlatformFieldSymbol(
     override val identifier: Identifier,
     val ofTypeSymbol: BasicTypeSymbol
 ) : NamedSymbolTableElement()
-
-/**
- * Data Types
- */
-data class PlatformObjectSymbol(
-    val identifier: Identifier,
-    val featureSupport: FeatureSupport,
-    private val symbolTable: SymbolTable = SymbolTable(NullSymbolTable)
-) : Symbol(), Type, Scope<Symbol> by symbolTable
-
-data class ObjectSymbol(
-    val qualifiedName: String,
-    val identifier: Identifier,
-    val featureSupport: FeatureSupport
-) : Symbol(), Type
-
-data class GroundRecordTypeSymbol(
-    override val parent: Scope<Symbol>,
-    val qualifiedName: String,
-    override val identifier: Identifier,
-    val featureSupport: FeatureSupport
-) : NamedSymbolWithMembers(parent), Type {
-    lateinit var fields: List<FieldSymbol>
-}
-
-data class ParameterizedRecordTypeSymbol(
-    override val parent: Scope<Symbol>,
-    val qualifiedName: String,
-    override val identifier: Identifier,
-    val featureSupport: FeatureSupport
-) : NamedSymbolWithMembers(parent), RawTerminus, Type {
-    override lateinit var typeParams: List<TypeParameter>
-    lateinit var fields: List<FieldSymbol>
-}
-
-/**
- * Basic Types
- */
-data class BasicTypeSymbol(
-    val identifier: Identifier,
-    private val symbolTable: SymbolTable = SymbolTable(NullSymbolTable)
-) : Symbol(), Type, Scope<Symbol> by symbolTable
-
-data class ParameterizedBasicTypeSymbol(
-    val identifier: Identifier,
-    val instantiation: SingleTypeInstantiation,
-    val featureSupport: FeatureSupport,
-    private val symbolTable: SymbolTable = SymbolTable(NullSymbolTable)
-) : Symbol(), RawTerminus, Type, Scope<Symbol> by symbolTable {
-    override lateinit var typeParams: List<TypeParameter>
-    lateinit var modeSelector: (List<Type>) -> BasicTypeMode
-    lateinit var fields: List<PlatformFieldSymbol>
-}
 
 /**
  * Plugins
