+++ src/main/kotlin/org/shardscript/semantics/core/Type.kt
@@ -1,0 +1,119 @@
+package org.shardscript.semantics.core
+
+import org.shardscript.semantics.infer.SubstitutionChain
+
+sealed interface Type
+
+data object ErrorSymbol : Symbol(), Type
+
+data class FunctionTypeSymbol(
+    val formalParamTypes: List<Type>,
+    val returnType: Type
+) : Symbol(), Type
+
+sealed class TypeParameter : Symbol(), Type
+
+data class StandardTypeParameter(
+    val qualifiedName: String,
+    val identifier: Identifier
+) : TypeParameter(), Type
+
+data class ImmutableFinTypeParameter(
+    val qualifiedName: String,
+    val identifier: Identifier
+) : TypeParameter(), CostExpression, Type {
+    override fun <R> accept(visitor: CostExpressionVisitor<R>): R {
+        return visitor.visit(this)
+    }
+}
+
+data class MutableFinTypeParameter(
+    val qualifiedName: String,
+    val identifier: Identifier
+) : TypeParameter(), CostExpression, Type {
+    override fun <R> accept(visitor: CostExpressionVisitor<R>): R {
+        return visitor.visit(this)
+    }
+}
+
+data class FinTypeSymbol(val magnitude: Long) : Symbol(), CostExpression, Type {
+    override fun <R> accept(visitor: CostExpressionVisitor<R>): R {
+        return visitor.visit(this)
+    }
+}
+
+data object ConstantFinTypeSymbol : Symbol(), CostExpression, Type {
+    override fun <R> accept(visitor: CostExpressionVisitor<R>): R {
+        return visitor.visit(this)
+    }
+}
+
+data class SumCostExpression(val children: List<CostExpression>) : Symbol(), CostExpression {
+    override fun <R> accept(visitor: CostExpressionVisitor<R>): R {
+        return visitor.visit(this)
+    }
+}
+
+data class ProductCostExpression(val children: List<CostExpression>) : Symbol(), CostExpression {
+    override fun <R> accept(visitor: CostExpressionVisitor<R>): R {
+        return visitor.visit(this)
+    }
+}
+
+data class MaxCostExpression(val children: List<CostExpression>) : Symbol(), CostExpression {
+    override fun <R> accept(visitor: CostExpressionVisitor<R>): R {
+        return visitor.visit(this)
+    }
+}
+
+data class SymbolInstantiation(
+    val substitutionChain: SubstitutionChain
+) : Symbol(), Type
+
+data class PlatformObjectSymbol(
+    val identifier: Identifier,
+    val featureSupport: FeatureSupport,
+    private val symbolTable: SymbolTable = SymbolTable(NullSymbolTable)
+) : Symbol(), Type, Scope<Symbol> by symbolTable
+
+data class ObjectSymbol(
+    val qualifiedName: String,
+    val identifier: Identifier,
+    val featureSupport: FeatureSupport
+) : Symbol(), Type
+
+data class GroundRecordTypeSymbol(
+    override val parent: Scope<Symbol>,
+    val qualifiedName: String,
+    override val identifier: Identifier,
+    val featureSupport: FeatureSupport
+) : NamedSymbolWithMembers(parent), Type {
+    lateinit var fields: List<FieldSymbol>
+}
+
+data class ParameterizedRecordTypeSymbol(
+    override val parent: Scope<Symbol>,
+    val qualifiedName: String,
+    override val identifier: Identifier,
+    val featureSupport: FeatureSupport
+) : NamedSymbolWithMembers(parent), RawTerminus, Type {
+    override lateinit var typeParams: List<TypeParameter>
+    lateinit var fields: List<FieldSymbol>
+}
+
+data class BasicTypeSymbol(
+    val identifier: Identifier,
+    private val symbolTable: SymbolTable = SymbolTable(NullSymbolTable)
+) : Symbol(), Type, Scope<Symbol> by symbolTable
+
+data class ParameterizedBasicTypeSymbol(
+    val identifier: Identifier,
+    val instantiation: SingleTypeInstantiation,
+    val featureSupport: FeatureSupport,
+    private val symbolTable: SymbolTable = SymbolTable(NullSymbolTable)
+) : Symbol(), RawTerminus, Type, Scope<Symbol> by symbolTable {
+    override lateinit var typeParams: List<TypeParameter>
+    lateinit var modeSelector: (List<Type>) -> BasicTypeMode
+    lateinit var fields: List<PlatformFieldSymbol>
+}
+
